requirements.txt
allure-behave==2.13.5
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.0.0
astor==0.8.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
behave @ git+https://github.com/behave/behave@0d0203dce45038dcb05accb03cd3a2b49b3ff243
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
cucumber-tag-expressions==6.1.0
docutils==0.20.1
et-xmlfile==1.1.0
execnet==2.1.1
flaky==3.8.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
natsort==8.4.0
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
paramiko==3.4.0
parse==1.20.1
parse-type==0.6.2
path==16.14.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.17.2
PyNaCl==1.5.0
PySocks==1.7.1
pytest==8.2.2
pytest-ast-transformer==1.0.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==14.0
pytest-soft-assertions==0.1.2
pytest-xdist==3.6.1
requests==2.31.0
robotframework==7.0
robotframework-datadriver==1.11.0
robotframework-pabot==2.18.0
robotframework-pythonlibcore==4.3.0
robotframework-requests==0.9.6
robotframework-seleniumlibrary==6.2.0
robotframework-stacktrace==0.4.1
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
texttable==1.7.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
